Ajout d'une classe 'tableau' pour remplacer 'std::vector'.
Le but est de pouvoir calculer la mémoire utilisée.

Pour le moment la classe ne fait qu'enrober un std::vector avec une
logeuse_guardee spécialisée. Une implémentation spécifique du tableau
reste à finir pour ne pas subir trop d'overhead lors des appels de
fonctions.

Dans le future, d'autres conteneurs seront converti de la même manière.

Cette implémentation réduit également le nombre de transtypage, car des
'long' sont utilisés pour les index maintenant.

+++ mikisa/coeur/corps/volume.cc
@@ -24,7 +24,7 @@
 
 #include "volume.hh"
 
-#include "../logeuse_memoire.hh"
+#include "bibloc/logeuse_memoire.hh"
 
 /* ************************************************************************** */
 
@@ -33,9 +33,9 @@ type_primitive Volume::type_prim() const
 	return type_primitive::VOLUME;
 }
 
-size_t BaseGrille::calcul_index(size_t x, size_t y, size_t z) const
+long BaseGrille::calcul_index(size_t x, size_t y, size_t z) const
 {
-	return x + (y + z * m_res[1]) * m_res[0];
+	return static_cast<long>(x + (y + z * m_res[1]) * m_res[0]);
 }
 
 bool BaseGrille::hors_des_limites(size_t x, size_t y, size_t z) const
